requirements.txt
assemblyai
"assemblyai[extras]"
deepgram-sdk
